refactor: relocate hrp tests
