+++ hrp/tests/compat_test.go
@@ -1,4 +1,4 @@
-package examples
+package tests
 
 import (
 	"testing"
@@ -8,8 +8,8 @@ import (
 
 // generated by examples/data/har/demo.har using HttpRunner v3.1.6
 var (
-	demoHttpRunnerJSONPath hrp.TestCasePath = "demo_httprunner.json"
-	demoHttpRunnerYAMLPath hrp.TestCasePath = "demo_httprunner.yaml"
+	demoHttpRunnerJSONPath hrp.TestCasePath = "../../examples/hrp/demo_httprunner.json"
+	demoHttpRunnerYAMLPath hrp.TestCasePath = "../../examples/hrp/demo_httprunner.yaml"
 )
 
 func TestCompatTestCase(t *testing.T) {
